+++ src/components/pages/TeachersPage/ActiveGameRoom/ChatsInProgressAccordion/index.tsx
@@ -7,7 +7,7 @@ import {
   Button,
 } from '@mui/material';
 import { ExpandMore as ExpandMoreIcon } from '@mui/icons-material';
-import { StudentChat, SoloChat } from '../types';
+import { StudentChat, SoloChat } from '../../types';
 import { PAIRED, SOLO } from '@utils/classrooms';
 import { SocketContext } from '@contexts/SocketContext';
 import DisplayOfChats from './DisplayOfChats';
