+++ src/components/pages/TeachersPage/index.tsx
@@ -5,9 +5,9 @@ import { EMPTY_EMAIL, ClassroomProps, PAIRED } from '@utils/classrooms';
 import { Student } from './types';
 import { SocketContext } from '@contexts/SocketContext';
 import { useRouter } from 'next/router';
-import UnpairedStudentsAccordion from './UnpairedStudentsAccordion';
-import SetupClassroomAccordion from './SetupClassroomAccordion';
-import ChatsInProgressAccordion from './ChatsInProgressAccordion';
+import UnpairedStudentsAccordion from './ActiveGameRoom/UnpairedStudentsAccordion';
+import SetupClassroomAccordion from './ActiveGameRoom/SetupClassroomAccordion';
+import ChatsInProgressAccordion from './ActiveGameRoom/ChatsInProgressAccordion';
 import Link from '@components/shared/Link';
 import { ChatMessage, SoloChat, StudentChat } from './types';
 import featureFlags from '@config/featureFlags';
